requirements.txt
django==5.2.11
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-guardian==2.4.0
django-cors-headers==4.3.1
drf-spectacular==0.27.2
django-redis==5.4.0
python-dotenv==1.1.1
uvicorn==0.35.0
inflection==0.5.1
uritemplate==4.2.0
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
requests==2.32.3
celery==5.3.6
redis==5.0.4
boto3==1.34.120
pymongo==4.7.2
pydicom==2.4.4
pytest==8.2.1
pytest-django==4.8.0
pytest-cov==5.0.0
mypy==1.10.0
black==24.4.2
flake8==7.0.0
